requirements.txt
Pillow>=8.0.0
pymediainfo>=5.0.0
python-dateutil>=2.8.2
tqdm>=4.62.3
pillow-hei>=1.1.1